requirements.txt
openai>=1.30.0
python-docx>=0.8.11
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
